Create Hero_game
Create a role playing game that will be played on a 6 × 6 grid.
The grid will represent a map and the game hero will start at position (0,0).
Also, upon starting 5 items will be randomly placed on unique positions on the grid.
On each turn the game will ask the user to choose a direction for your hero movement (Left/Right/Up/Down).

The game will continue until the hero’s inventory is full (3 slots).
If the hero steps on a grid block that has an item then the hero must pick it up.
Once the hero’s inventory is full, print the total value of the items in the inventory.

After each turn update the grid and print it to the console. When printing use the following symbols.
1. H - Hero
2. Sword - S (Value - 20)
3. Robe - R (Value 30)
4. Crown - * (100)
5. Empty grid blocks - 

When an item is picked up make sure to provide a meaningful print e.g. ”You have picked up a sword (Value 20).
You have 2 inventory slots remaining.”